requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
pytest==4.4.1
requests==2.21.0
Flask-SQLAlchemy==2.3.2
python-dotenv==0.10.3
Flask-Migrate==2.5.2
freezegun